Switch operatori 2

+++ Program.cs
@@ -49,12 +49,12 @@
         //                 : "1st and 2nd numbers are not negative";
         // System.Console.WriteLine(javob3);
                 System.Console.Write("Birinchi sonni kiriting: ");
-        firstNumber = Convert.ToInt32 (Console .ReadLine());
+        firstNumber = Convert.ToDecimal (Console .ReadLine());
         System.Console.WriteLine("Qanday amal bajaray  + , - , / , * ");
          string operation = Console.ReadLine();
 
         System.Console.Write("Ikkinchi sonni kiriring: ");
-        secondNumber = Convert.ToInt32(Console .ReadLine());
+        secondNumber = Convert.ToDecimal(Console .ReadLine());
                switch (operation){
                 case "+" : 
                         System.Console.WriteLine($"{firstNumber} + {secondNumber} = {firstNumber + secondNumber}"); break ;
@@ -69,3 +69,4 @@
         }
 
         
+
